Bug 1067998 - Fix OOM crash in gfxAlphaBoxBlur::Init on large blur surface. r=Bas
UltraBlame original commit: 56aa6f15146787d44935f767aa87d0cc22ddc5e9

+++ gfx/2d/Blur.h
@@ -99,7 +99,7 @@ public:
 
 
 
-  int32_t GetSurfaceAllocationSize() const;
+  size_t GetSurfaceAllocationSize() const;
 
   
 
@@ -162,7 +162,7 @@ private:
   
 
 
-  int32_t mSurfaceAllocationSize;
+  size_t mSurfaceAllocationSize;
